Hi Don, circling back to this at long last.  Kim approved all the changes proposed by Jaron as non-substantive, so we're ready to incorporate them.  Michael did just have one comment:
Amanda, the URL at the end of Communications and Publications doesn't look right to me.  I think it will show up twice the way it is currently written.  Will someone who knows LaTeX be submitting these changes?  They will know what to do if they are aware of the need to fix it.

Meagan and I still have to make a couple more board-approved changes that should be staightforward--but would you be able to get us to the point everything else is cleaned up, as one of your last (gracious!) acts on the Faculty Handbook?

AS
--
Dr. Amanda M. Sparkman